+++ dynamo/prediction/perturbation.py
@@ -10,12 +10,10 @@ from ..vectorfield import SvcVectorField
 from ..vectorfield.scVectorField import vector_field_function_knockout
 from ..vectorfield.vector_calculus import (
     jacobian,
-    rank_cell_groups,
-    rank_cells,
-    rank_genes,
     vecfld_from_adata,
     vector_transformation,
 )
+from ..vectorfield.rank_vf import rank_cell_groups, rank_cells, rank_genes
 from .utils import expr_to_pca, pca_to_expr, z_score, z_score_inv
