+++ dynamo/vectorfield/stochastic_process.py
@@ -3,11 +3,10 @@ from sklearn.neighbors import NearestNeighbors
 from tqdm import tqdm
 from anndata import AnnData
 from typing import Optional, List
-from scVectorField import VecFldDict
 
 from ..tools.connectivity import _gen_neighbor_keys, check_and_recompute_neighbors
 from ..tools.utils import log1p_
-from .utils import vecfld_from_adata, vector_field_function
+from .utils import VecFldDict, vecfld_from_adata, vector_field_function
 
 
 def diffusionMatrix(
